refactor: command has been renamed to db:capsule
